backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
pydantic-settings==2.5.2
python-multipart==0.0.9
psycopg[binary]==3.2.1
